+++ compages/__init__.py
@@ -5,8 +5,8 @@ from ._structure import (
     StructuringError,
 )
 from ._structure_handlers import (
-    StructureDictIntoDataclass,
-    StructureDictIntoNamedTuple,
+    StructureMappingIntoDataclass,
+    StructureMappingIntoNamedTuple,
     simple_structure,
     structure_into_bool,
     structure_into_bytes,
@@ -18,8 +18,8 @@ from ._structure_handlers import (
     structure_into_str,
     structure_into_tuple,
     structure_into_union,
-    structure_list_into_dataclass,
-    structure_list_into_named_tuple,
+    structure_sequence_into_dataclass,
+    structure_sequence_into_named_tuple,
 )
 from ._unstructure import (
     Unstructurer,
@@ -48,8 +48,8 @@ from ._unstructure_handlers import (
 __all__ = [
     "DataclassBase",
     "NamedTupleBase",
-    "StructureDictIntoDataclass",
-    "StructureDictIntoNamedTuple",
+    "StructureMappingIntoDataclass",
+    "StructureMappingIntoNamedTuple",
     "Structurer",
     "StructurerContext",
     "StructuringError",
@@ -71,8 +71,8 @@ __all__ = [
     "structure_into_str",
     "structure_into_tuple",
     "structure_into_union",
-    "structure_list_into_dataclass",
-    "structure_list_into_named_tuple",
+    "structure_sequence_into_dataclass",
+    "structure_sequence_into_named_tuple",
     "unstructure_as_bool",
     "unstructure_as_bytes",
     "unstructure_as_dict",